requirements.txt
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
pillow==10.4.0
sympy==1.13.1
torch==2.4.0
typing_extensions==4.12.2
